# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "operationId": "Categories_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "true = เฉพาะที่เปิดใช้งาน (ใช้ในฟอร์มแจ้งซ่อม)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "รายการหมวดหมู่งานซ่อม",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "Categories_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CategoryDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "เพิ่มหมวดหมู่",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "Categories_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CategoryDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ข้อมูลหมวดหมู่",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "Categories_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CategoryDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "แก้ไขหมวดหมู่ / เปิด-ปิดการใช้งาน",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "Categories_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD_REQUEST / VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — ไม่มี token หรือ token ใช้ไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — สิทธิ์ไม่พอ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT — ชนกฎธุรกิจ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ลบหมวดหมู่ที่ยังไม่มีใบแจ้งซ่อม",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "csmju-maintenance-request API",
    "description": "ระบบแจ้งซ่อม — ระบบย่อยของ CSMJU2030 · ทุก endpoint ใต้ /api/v1 ต้องมี Core Hub access token (Authorization: Bearer หรือคุกกี้ csmju_maintenance_request_access_token) · response ห่อด้วย envelope มาตรฐาน",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT (RS256 จาก Core Hub)",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorCode": {
        "type": "string",
        "enum": [
          "BAD_REQUEST",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "CONFLICT",
          "TOO_MANY_REQUESTS",
          "INTERNAL_ERROR",
          "SERVICE_UNAVAILABLE"
        ]
      },
      "ErrorBodyDto": {
        "type": "object",
        "properties": {
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              }
            ]
          },
          "message": {
            "type": "string",
            "description": "ข้อความภาษาไทยที่แสดงให้ผู้ใช้ได้ทันที"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBodyDto"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "DeletedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "CategoryIcon": {
        "type": "string",
        "enum": [
          "computer",
          "monitor",
          "projector",
          "aircon",
          "fan",
          "light",
          "network",
          "audio",
          "furniture",
          "other"
        ]
      },
...
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "คอมพิวเตอร์"
          },
          "symptoms": {
            "example": [
              "เปิดไม่ติด",
              "จอไม่แสดงภาพ"
            ],
            "description": "ปุ่มอาการในหน้าแจ้งซ่อม",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean",
            "description": "false = ไม่แสดงในฟอร์มแจ้งซ่อม แต่ข้อมูลเดิมยังอ้างถึงได้"
          },
          "requestCount": {
            "type": "number",
            "description": "จำนวนใบแจ้งซ่อมจริงในหมวดนี้"
          },
          "openRequestCount": {
            "type": "number",
            "description": "ใบที่ยังไม่ปิดงาน"
          },
          "equipmentCount": {
            "type": "number",
            "description": "จำนวนเครื่องในห้องต่าง ๆ ที่เป็นประเภทนี้"
          }
        },
        "required": [
          "id",
          "name",
          "symptoms",
          "icon",
          "sortOrder",
          "isActive",
          "requestCount",
          "openRequestCount",
          "equipmentCount"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "คอมพิวเตอร์"
          },
          "symptoms": {
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "คอมพิวเตอร์"
          },
          "symptoms": {
            "maxItems": 12,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "icon": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryIcon"
              }
            ]
          },
          "sortOrder": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}